+++ libs/application/templates/social-insurance-administration/additional-support-for-the-elderly/src/fields/Review/review-groups/Payment.tsx
@@ -6,13 +6,12 @@ import {
 } from '@island.is/application/ui-components'
 import { GridColumn, GridRow } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { additionalSupportForTheElderyFormMessage } from '../../../lib/messages'
 import { socialInsuranceAdministrationMessage } from '@island.is/application/templates/social-insurance-administration-core/messages'
 import { ReviewGroupProps } from './props'
 import { getApplicationAnswers } from '../../../lib/additionalSupportForTheElderlyUtils'
 import { BankAccountType } from '@island.is/application/templates/social-insurance-administration-core/constants'
 import { YES } from '@island.is/application/types'
-import { getTaxLevelOption } from './utils'
+import { getTaxLevelOption } from '@island.is/application/templates/social-insurance-administration-core/socialInsuranceAdministrationUtils'
 
 export const Payment = ({
   application,
@@ -122,7 +121,7 @@ export const Payment = ({
         >
           <RadioValue
             label={formatMessage(
-              additionalSupportForTheElderyFormMessage.review.personalAllowance,
+              socialInsuranceAdministrationMessage.confirm.personalAllowance,
             )}
             value={personalAllowance}
           />
@@ -135,7 +134,7 @@ export const Payment = ({
           >
             <DataValue
               label={formatMessage(
-                additionalSupportForTheElderyFormMessage.review.ratio,
+                socialInsuranceAdministrationMessage.confirm.ratio,
               )}
               value={`${personalAllowanceUsage}%`}
             />
@@ -147,7 +146,7 @@ export const Payment = ({
         <GridColumn span={['12/12', '12/12', '12/12', '12/12']}>
           <DataValue
             label={formatMessage(
-              additionalSupportForTheElderyFormMessage.review.taxLevel,
+              socialInsuranceAdministrationMessage.payment.taxLevel,
             )}
             value={formatMessage(getTaxLevelOption(taxLevel))}
           />
