+++ libs/application/templates/social-insurance-administration/additional-support-for-the-elderly/src/lib/additionalSupportForTheElderlyUtils.ts
@@ -1,14 +1,8 @@
 import { getValueViaPath } from '@island.is/application/core'
-import {
-  Application,
-  Option,
-  NO,
-  YES,
-  YesOrNo,
-} from '@island.is/application/types'
+import { Application, YesOrNo } from '@island.is/application/types'
 import addMonths from 'date-fns/addMonths'
 import subMonths from 'date-fns/subMonths'
-import { AttachmentLabel, TaxLevelOptions } from './constants'
+import { AttachmentLabel } from './constants'
 import {
   FileType,
   Attachments,
@@ -16,8 +10,10 @@ import {
   BankInfo,
   PaymentInfo,
 } from '@island.is/application/templates/social-insurance-administration-core/types'
-import { BankAccountType } from '@island.is/application/templates/social-insurance-administration-core/constants'
-import { additionalSupportForTheElderyFormMessage } from './messages'
+import {
+  BankAccountType,
+  TaxLevelOptions,
+} from '@island.is/application/templates/social-insurance-administration-core/constants'
 
 enum AttachmentTypes {
   ADDITIONAL_DOCUMENTS = 'additionalDocuments',
@@ -196,25 +192,6 @@ export function getAttachments(application: Application) {
   }
 
   return attachments
-}
-
-export function getTaxOptions() {
-  const options: Option[] = [
-    {
-      value: TaxLevelOptions.INCOME,
-      label: additionalSupportForTheElderyFormMessage.payment.taxIncomeLevel,
-    },
-    {
-      value: TaxLevelOptions.FIRST_LEVEL,
-      label: additionalSupportForTheElderyFormMessage.payment.taxFirstLevel,
-    },
-    {
-      value: TaxLevelOptions.SECOND_LEVEL,
-      label: additionalSupportForTheElderyFormMessage.payment.taxSecondLevel,
-    },
-  ]
-
-  return options
 }
 
 // returns available years. Available period is
