+++ libs/application/templates/social-insurance-administration/additional-support-for-the-elderly/src/lib/constants.ts
@@ -1,12 +1,6 @@
 import { socialInsuranceAdministrationMessage } from '@island.is/application/templates/social-insurance-administration-core/messages'
 import { MessageDescriptor } from 'react-intl'
 
-export enum TaxLevelOptions {
-  INCOME = '2',
-  FIRST_LEVEL = '1',
-  SECOND_LEVEL = '3',
-}
-
 export const AttachmentLabel: {
   [key: string]: MessageDescriptor
 } = {
