+++ libs/application/templates/social-insurance-administration/old-age-pension/src/lib/dataSchema.ts
@@ -1,6 +1,6 @@
 import { parsePhoneNumberFromString } from 'libphonenumber-js'
 import { z } from 'zod'
-import { ApplicationType, TaxLevelOptions, Employment } from './constants'
+import { ApplicationType, Employment } from './constants'
 import {
   formatBankInfo,
   validIBAN,
@@ -8,7 +8,10 @@ import {
 } from '@island.is/application/templates/social-insurance-administration-core/socialInsuranceAdministrationUtils'
 import { NO, YES } from '@island.is/application/types'
 import { errorMessages } from '@island.is/application/templates/social-insurance-administration-core/messages'
-import { BankAccountType } from '@island.is/application/templates/social-insurance-administration-core/constants'
+import {
+  BankAccountType,
+  TaxLevelOptions,
+} from '@island.is/application/templates/social-insurance-administration-core/constants'
 
 const isValidPhoneNumber = (phoneNumber: string) => {
   const phone = parsePhoneNumberFromString(phoneNumber, 'IS')
