+++ libs/application/templates/social-insurance-administration/old-age-pension/src/lib/oldAgePensionUtils.ts
@@ -6,17 +6,14 @@ import {
   earlyRetirementMaxAge,
   ApplicationType,
   Employment,
-  TaxLevelOptions,
   AttachmentLabel,
   AttachmentTypes,
 } from './constants'
 import {
-  Option,
   Application,
   NationalRegistryResidenceHistory,
   YesOrNo,
 } from '@island.is/application/types'
-import { oldAgePensionFormMessage } from './messages'
 
 import * as kennitala from 'kennitala'
 import addYears from 'date-fns/addYears'
@@ -32,6 +29,7 @@ import {
 import {
   BankAccountType,
   MONTHS,
+  TaxLevelOptions,
 } from '@island.is/application/templates/social-insurance-administration-core/constants'
 import {
   Attachments,
@@ -523,25 +521,6 @@ export function getCombinedResidenceHistory(
   })
 
   return [...combinedResidenceHistory].reverse()
-}
-
-export function getTaxOptions() {
-  const options: Option[] = [
-    {
-      value: TaxLevelOptions.INCOME,
-      label: oldAgePensionFormMessage.payment.taxIncomeLevel,
-    },
-    {
-      value: TaxLevelOptions.FIRST_LEVEL,
-      label: oldAgePensionFormMessage.payment.taxFirstLevel,
-    },
-    {
-      value: TaxLevelOptions.SECOND_LEVEL,
-      label: oldAgePensionFormMessage.payment.taxSecondLevel,
-    },
-  ]
-
-  return options
 }
 
 export function isMoreThan2Year(answers: Application['answers']) {
